test(selectors): Update snapshots for seclector changes
