+++ lib/containers/modifications-map.js
@@ -19,7 +19,7 @@ function mapStateToProps (state, props) {
     feedsById: state.project.feedsById,
     mapState: state.mapState,
     modificationsOnMap: select.modificationsOnMap(state, props),
-    qualifiedStops: select.fullyQualifiedRouteStops(state, props)
+    qualifiedStops: select.qualifiedStopsFromSegments(state, props)
   }
 }
