+++ lib/containers/reroute.js
@@ -11,7 +11,7 @@ function mapStateToProps (
   return {
     feeds: select.feedsWithBundleNames(state, ownProps),
     mapState: state.mapState,
-    qualifiedStops: select.fullyQualifiedRouteStops(state, ownProps),
+    qualifiedStops: select.qualifiedStopsFromSegments(state, ownProps),
     routePatterns: select.routePatterns(state, ownProps),
     segmentDistances: select.segmentDistances(state, ownProps),
     selectedFeed: select.modificationFeed(state, ownProps),
