+++ lib/modules/opportunity-datasets/components/list.js
@@ -10,15 +10,7 @@ import messages from '../../../utils/messages'
 
 import {Button} from '../../../components/buttons'
 import {Group} from '../../../components/input'
-import {
-  addOpportunityComponent,
-  checkUploadStatus,
-  clearStatus,
-  createGoTo,
-  deleteOpportunityDataset,
-  downloadOpportunityDataset,
-  removeOpportunityComponent
-} from '../actions'
+import * as actions from '../actions'
 import * as select from '../selectors'
 import Selector from './selector'
 
@@ -184,11 +176,6 @@ const Status = (status: UploadStatus & {clear: () => void}) =>
   </div>
 
 export default connect(createStructuredSelector(select), {
-  addOpportunityComponent,
-  checkUploadStatus,
-  clearStatus,
-  deleteOpportunityDataset,
-  downloadOpportunityDataset,
-  goToUpload: createGoTo('/upload'),
-  removeOpportunityComponent
+  ...actions,
+  goToUpload: actions.createGoTo('/upload')
 })(ListOpportunityDatasets)
