+++ lib/modules/opportunity-datasets/selectors.js
@@ -4,7 +4,8 @@ import sortBy from 'lodash/sortBy'
 import {createSelector} from 'reselect'
 
 import colors from '../../constants/colors'
-import * as select from '../../selectors'
+
+import selectCurrentRegionId from '../../selectors/current-region-id'
 
 import type {State} from './types'
 
@@ -38,7 +39,7 @@ export const activeOpportunityDatasetGrid = createSelector(
  * management
  */
 export const opportunitiesUrl = createSelector(
-  select.currentRegionId,
+  selectCurrentRegionId,
   (regionId) => `/regions/${regionId}/opportunities`
 )
 
@@ -61,7 +62,7 @@ export const activeOpportunityDatasetColorizer = createSelector(
 /**
  * Get all of the dataset upload statuses
  */
-export const selectUploadStatuses = (state: {
+export const uploadStatuses = (state: {
   opportunityDatasets: State
 }) =>
   state.opportunityDatasets.uploadStatuses
