+++ lib/modules/r5-version/__tests__/selectors.js
@@ -11,18 +11,6 @@ const TEST_VERSIONS = [
 ]
 
 describe('R5 Version > Selectors', () => {
-  it('should properly convert versions to numbers', () => {
-    ;[{
-      input: TEST_VERSIONS[2],
-      output: 20000
-    }, {
-      input: TEST_VERSIONS[0],
-      output: 30609.46
-    }].forEach(v => expect(select.versionToNumber(v.input)).toBe(v.output))
-
-    expect(select.versionToNumber('v3.2.1-1') > select.versionToNumber('v3.2.1')).toBeTruthy()
-  })
-
   it('should properly select and sort all valid versions', () => {
     const validVersions = select.allValidVersions({
       r5Version: {
