+++ lib/modules/r5-version/components/selector.js
@@ -18,7 +18,7 @@ import * as select from '../selectors'
 import type {ReactSelectResult} from '../../../types'
 
 type Props = {
-  allVersions: string[],
+  allValidVersions: string[],
   analysisVersion?: string,
   currentR5Version: string,
   newerVersionIsAvailable: boolean,
@@ -61,14 +61,14 @@ export class SelectR5Version extends Component<void, Props, State> {
   render () {
     const {showAllVersions} = this.state
     const {
-      allVersions,
+      allValidVersions,
       analysisVersion,
       currentR5Version,
       newerVersionIsAvailable,
       releaseVersions
     } = this.props
 
-    const versions = showAllVersions ? allVersions : releaseVersions
+    const versions = showAllVersions ? allValidVersions : releaseVersions
     const options = versions.map(v => {
       return {
         value: v,
@@ -78,7 +78,7 @@ export class SelectR5Version extends Component<void, Props, State> {
 
     // if this region has an older version that is no longer supported, show it
     // but strike it through
-    if (!allVersions.includes(currentR5Version)) {
+    if (!allValidVersions.includes(currentR5Version)) {
       options.push({
         value: currentR5Version,
         label: (
