+++ lib/modules/r5-version/selectors.js
@@ -6,10 +6,10 @@ import {createSelector} from 'reselect'
 
 import {
   MINIMUM_R5_VERSION,
-  RELEASE_VERSION_REGEX,
-  VERSION_PARSE_REGEX
+  RELEASE_VERSION_REGEX
 } from './constants'
 import {initialState} from './reducer'
+import {versionToNumber} from './utils'
 
 import type {State} from './types'
 
@@ -18,24 +18,6 @@ type FullState = {
 }
 
 const MIN_VERSION_NUM = versionToNumber(MINIMUM_R5_VERSION)
-
-/**
- * Convert an r5 version to a number for comparing
- */
-export function versionToNumber (version: string): number {
-  const versionMatch = version.match(VERSION_PARSE_REGEX)
-  if (versionMatch) {
-    const [, major, minor, patch, commit] = versionMatch
-
-    // NB Assumes no major, minor, patch, or commit version exceeds 100
-    return parseInt(major) * 10000 +
-      parseInt(minor) * 100 +
-      parseInt(patch) +
-      parseInt(commit || 0) / 100
-  }
-
-  return -1
-}
 
 /**
  * Use this to set the state key
