+++ lib/selectors/index.js
@@ -34,6 +34,7 @@ export {default as fullyQualifiedRouteStops} from './fully-qualified-route-stops
 export {default as isochrone} from './isochrone'
 export {default as lastStopDistanceFromStart} from './last-stop-distance-from-start'
 export {default as maxAccessibility} from './max-accessibility'
+export {default as maxTripDurationMinutes} from './max-trip-duration-minutes'
 export {default as modificationBounds} from './modification-bounds'
 export {default as modificationFeedIsLoaded} from './modification-feed-is-loaded'
 export {default as modificationFeed} from './modification-feed'
@@ -46,6 +47,7 @@ export {default as modificationsWithTimetables} from './modifications-with-timet
 export {default as modifications} from './modifications'
 export {default as numberOfStops} from './number-of-stops'
 export {default as percentileCurves} from './percentile-curves'
+export {default as profileRequestHasChanged} from './profile-request-has-changed'
 export {default as projectIsLoaded} from './project-is-loaded'
 export {default as projectTimetables} from './project-timetables'
 export {default as qualifiedStopsFromSegments} from './qualified-stops-from-segments'
